Tweak ")"'s behaviour in paredit
If I type ") ", then I don't want to be ")\n " inserted. So we write
our own replacement for `paredit-close-round-and-newline` that sets a
transient map with a single entry on `SPC` which calls
(just-one-space -1).

TODO: A <backspace> binding that removes the inserted newline and
indentation and puts point after the closed paren