Start cutting over to use-package
req-package is proving to be very annoying in terms of working out
whether/when a package will actually get loaded, and the source is
sufficiently convoluted that I can't work out whether the bug is in
req-package or my understanding of it. I think I can cope with using
'use-package' directly.