Support capturing a week summary for weeknotes
Right now, it's not quite right. Really it needs to find the most recent OPEN
weeknote and select that or, if we're in a new week and there aren't any open
notes, to create a new weeknote and jump to its summary.

However, this (sort of) works and it's always good to have a shitty first draft
in place.